[REM-103] feat: implement RunnerServices and MenuManager for server management
- crawjud/context_manager/runner.py: added RunnerServices class to manage server components in separate threads
- crawjud/context_manager/menu.py: created MenuManager class for handling user interactions and menu options
- crawjud/__init__.py: integrated RunnerServices and MenuManager into MasterApp, added properties for Celery and Quart instances
